validacion de login y registro de prd

+++ lib/domain/entities/productos/productosEntity.dart
@@ -21,7 +21,7 @@ class ProductosEntity extends Equatable {
   String descripcion;
   int stock;
   double costo;
-  int precio;
+  double precio;
   int unidad;
   String estado;
   DateTime? fecha;
